Feature: Plate report storage in MongoDB; Plate path playback query API; Unique object ID querying
This version features the storage and querying of plate number paths. The
logic is similar to that of user paths, but the PlateData model supports
multiple reports per timestamp, so the returned payload is more complex
than that of user reports.

To enable rapid searching of user and plate objects in the database,
the MongoGetUniqueFieldValues() method has been introduced to quickly
query unique values for a certain key.

- Added QueryAvailableObjects() method for querying unique users/plates in DB
- Renamed QueryObjectPath to QueryUserPath()
- Added QueryPlatePath()
- Added MongoGetUniqueFieldValues() generic method to MongoDB service
- Added StorePlateReportToMongoDB() method to RTTMAS service
- Added /playback/objects API endpoint
- Temporarily disabled Host assertion in webserver.go